LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-handler-throws-expected.txt


FAIL event.intercept() should abort if the handler throws assert_equals: expected "#1" but got ""
